+++ src/containers/appRouter/index.tsx
@@ -1,7 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { Header } from 'components/common/Header'
 import { Footer } from 'components/common/Footer'
-import { Suspense } from 'react'
 
 import { Home } from 'containers/home/Home'
 
@@ -11,27 +10,24 @@ const routes = [
     exact: true,
     component: Home,
   },
+  // add more routes here
 ]
 
 export const AppRouter = () => {
-  const LoadingMessage = () => <div>Loading..,</div>
   return (
     <div>
       <Header />
       <BrowserRouter>
-        <Suspense fallback={<LoadingMessage />}>
-          <Routes>
-            {routes.map(route => (
-              <Route
-                path={route.path}
-                element={route.component()}
-                key={route.path}
-              />
-            ))}
-          </Routes>
-        </Suspense>
+        <Routes>
+          {routes.map(route => (
+            <Route
+              path={route.path}
+              element={route.component()}
+              key={route.path}
+            />
+          ))}
+        </Routes>
       </BrowserRouter>
-
       <Footer />
     </div>
   )
