+++ src/containers/home/styles.ts
@@ -1,0 +1,6 @@
+import { css} from '@emotion/css'
+
+export const styles = {
+  loading: css({ textAlign: 'center' }),
+  cardRow: css({ marginBottom: '0px !important' })
+}
